separated different sections of the code into modules. now we have experiments, results and the model itself as 3 separate modules.
