+++ baseline/experiments/base_size_experiment.py
@@ -4,11 +4,12 @@
 import csv
 
 # Ipmort the model and the generator
-from model import Trainer
-from generator import Generator
+from model.nn import Trainer
+from model.generator import Generator
 
 
-HIDDEN_SIZES = [100, 200, 400]
+BASE_SIZES = [16, 18, 20, 22, 24, 26, 28, 30, 32, 34, 36, 38, 40, 42]
+TRAINING_SIZES = [200000, 500000]
 ALPHABET = 'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmno'
 
 def train_model(base, training_size, length=2, choice=5, epochs=2000, batch_size=1000, dimension = 100, testing_size=100):
@@ -31,14 +32,15 @@ def train_model(base, training_size, length=2, choice=5, epochs=2000, batch_size
 
     return (base, train_acc, test_acc)
 
-def run():
-    results = []
-    for hidden_size in HIDDEN_SIZES:
-        result = train_model(base=36, training_size = 500000, dimension = hidden_size)
-        results.append(result)
-        with open("baseline/results/hidden_size_experiment/hidden_size_experiment_hidden_size_" + str(hidden_size) + ".csv", "w") as f:
+def run_experiment():
+    for training_size in TRAINING_SIZES:
+        results = []
+        for base_size in BASE_SIZES:
+            result = train_model(base=base_size, training_size = training_size)
+            results.append(result)
+        with open("baseline/results/base_size_experiment/base_size_experiment_training_size_" + str(training_size) + ".csv", "w") as f:
              writer = csv.writer(f)
              writer.writerows(results)
-
 if __name__ == '__main__':
-    run()
+    print('experimenting with base size')
+    run_experiment()
